Added definitions of \refbimodindex{} and \refstmodindex{}, which should be used to add index references for built-in and standard modules, respectively.
Modified \bimodindex{} and \stmodindex{} to make the page number bold, to
allow the defining instance of a module to stand out in the index.

Check-ins which fix improper use of \bimodindex{} and \stmodindex{} will be
made as fixes are applied.  Misc. indexing updates will occur as a side
effect in some cases.