Instead of bad hacks trying to worm around the inherited object.__reduce__, do a getattr() on the class so we can explicitly test for it.  The reduce()-calling code becomes a bit more regular as a result.
Also add support slots: if an object has slots, the default state is
(dict, slots) where dict is the __dict__ or None, and slots is a dict
mapping slot names to slot values.  We do a best-effort approach to
find slot names, assuming the __slots__ fields of classes aren't
modified after class definition time to misrepresent the actual list
of slots defined by a class.