Remove NEWS entries for packaging test suite fixes.
I haven’t updated NEWS for every change or fix in packaging, thinking
that the only entry we need for 3.3a1 is “Add packaging module to the
stdlib”.  Other developers have however added entries for these fixes,
but I don’t think they are interesting for users.

Note that the distutils2 backport has a full changelog.