whatsnew: tweak a couple importlib porting notes.
I had missed that there was already an entry for the Frozen package change, so
I combined mine with the existing one.  Also added the info about the reload
change to the entry for the other changes to reload, but I'm missing an issue
number for the rest of the attribute behavior changes so I left myself an
XXX as a reminder to find it.