bpo-34945: Buffer output in test suite only when creating junit file (GH-10204)
After commit dfe84af95fcd805f0667a3f4fd9d3003c2d675c1, the output of the
test suite is always buffered as the test output needs to be included in
the JUnit file in same cases (as when a test fails). This has the
consequence that printing or using debuggers (like pdb) in the test
suite does not result in a good user experience anymore.

This commit modifies the test suite runner so it only captures the test
output when the JUnit file is requested to fix the regression so prints
and debuggers are usable again.