📝 Make sure the phrase "constant-time compare" actually appears in the docs (GH-93396)
This is purely for SEO as this is the actual generic name for this kind of method and it currently does not appear in a Google search for "python constant time compare". Not creating an issue or setting this up for backports as its trivial (I think) and not a functional change.