gh-99942: python.pc on android/cygwin should link to libpython per configure.ac (GH-100356)
In commit 6c766b8826a23b71a59aff27c65efc77c9756e47 a previous change to avoid linking to libpython was partially reverted for Android (and later Cygwin as well), to add back the link flags. This was applied to distutils and to python-config.sh, but not to python.pc.

Add it back to python.pc as well.

Automerge-Triggered-By: GH:gpshead